Put sorting method in country adapter onbinndviewholder to account for the name changes made there instead of directly in study activity
